// src/enums/balance-type.ts
export enum BalanceType {
    FACTORING_WALLET = 'factoring_wallet',
    USHIPPER_WALLET = 'ushipper_wallet',
    DISPATCH_WALLET = 'dispatch_wallet',
    INTERNAL_USER_WALLET = 'internal_user_wallet',
    EXTERNAL_CARD_WALLET = 'external_card_wallet',
    EXTERNAL_BANK_WALLET = 'external_bank_wallet',
    COD_WALLET = 'cod_wallet',
    BROKER_WALLET = 'broker_wallet',
    CUSTOM_INTERNAL_WALLET = 'custom_internal_wallet',
}